Leon/homepage v2 (#9)
* add @types/uuid npm module

* hide v2 homepage view

* fixed chart component render issue

* homepage v2

* minor card styling
