+++ components/SingleExploreCard/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import Link from 'next/link';
 import { VerifiedIcon } from '../Shared/SvgIcons';
 import {
   Container,
@@ -28,10 +29,14 @@ export const SingleExploreCard = (props: SingleExplorecardProps) => {
           <Description>{card?.description}</Description>
           <ButtonWrapper>
             <button className='primary-btn'>
-              <span>View on PARAS</span>
+              <a href={card?.para_link} target="_blank" rel='noreferrer' style={{ color: "white" }}>
+                <span>View on PARAS</span>
+              </a>
             </button>
             <button className='primary-btn-naked'>
-              <span>Stake Your NFTs</span>
+              <Link href="/stake">
+                <span>Stake Your NFTs</span>
+              </Link>
             </button>
           </ButtonWrapper>
         </FooterWrapper>
