+++ components/Top30Ranked/index.tsx
@@ -1,4 +1,4 @@
-import React, { useRef } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import Carousel from 'react-multi-carousel';
 import 'react-multi-carousel/lib/styles.css';
 import { SingleExploreCard } from '../SingleExploreCard';
@@ -12,6 +12,8 @@ import {
   ArrowButton
 } from './styles';
 import { SingleRankCard } from '../SingleRankCard';
+import { getCollectionData, getCollectionStats, getHistoricalCollections, getTopCollections } from '../../utils/paraApi';
+import { parseEther } from '../../utils/bignumber';
 
 export const Top30Ranked = () => {
   const carouselRef = useRef<any>();
@@ -36,6 +38,29 @@ export const Top30Ranked = () => {
     }
   };
 
+  const [top30Collections, setTop30Collections] = useState<any[]>([]);
+
+  const getTop30Collections = async () => {
+
+    const _collections = await getHistoricalCollections();
+    console.log(_collections);
+    const collections = await getTopCollections();
+    const derivatives = await Promise.all(collections.map(async (data: any) => {
+      const { collection_id } = data;
+      const { floor_price, total_card_sale, volume } = await getCollectionStats(collection_id);
+      const { collection: name, media, socialMedia } = await getCollectionData(collection_id);
+      return {
+        name,
+        photo: `https://ipfs.io/ipfs/${media}`,
+        social_media: socialMedia,
+        floor_price: parseEther(floor_price),
+        total_listed: total_card_sale,
+        total_volume: parseEther(volume)
+      }
+    }))
+    setTop30Collections(derivatives);
+  }
+
   const goToNext = () => {
     const nextSlide = carouselRef.current.state.currentSlide + 1;
     carouselRef.current.goToSlide(nextSlide)
@@ -45,6 +70,10 @@ export const Top30Ranked = () => {
     const prevSlide = carouselRef.current.state.currentSlide - 1;
     carouselRef.current.goToSlide(prevSlide)
   }
+
+  useEffect(() => {
+    getTop30Collections();
+  }, []);
 
   return (
     <Container>
@@ -80,18 +109,35 @@ export const Top30Ranked = () => {
         >
           <div className='row'>
             <div className="col-md-4">
-              {[...Array(5).keys()].map(i => (
-                <SingleRankCard key={i} index={i + 1} />
+              {top30Collections.slice(0, 5).map((item, i) => (
+                <SingleRankCard key={i} index={i + 1} card={item} />
               ))}
             </div>
             <div className="col-md-4">
-              {[...Array(5).keys()].map(i => (
-                <SingleRankCard key={i} index={i + 6} />
+              {top30Collections.slice(5, 10).map((item, i) => (
+                <SingleRankCard key={i} index={i + 6} card={item} />
               ))}
             </div>
             <div className="col-md-4">
-              {[...Array(5).keys()].map(i => (
-                <SingleRankCard key={i} index={i + 11} />
+              {top30Collections.slice(10, 15).map((item, i) => (
+                <SingleRankCard key={i} index={i + 11} card={item} />
+              ))}
+            </div>
+          </div>
+          <div className='row'>
+            <div className="col-md-4">
+              {top30Collections.slice(15, 20).map((item, i) => (
+                <SingleRankCard key={i} index={i + 16} card={item} />
+              ))}
+            </div>
+            <div className="col-md-4">
+              {top30Collections.slice(20, 25).map((item, i) => (
+                <SingleRankCard key={i} index={i + 21} card={item} />
+              ))}
+            </div>
+            <div className="col-md-4">
+              {top30Collections.slice(25, 30).map((item, i) => (
+                <SingleRankCard key={i} index={i + 26} card={item} />
               ))}
             </div>
           </div>
