+++ components/TopMarketCap/index.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useState } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import Carousel from 'react-multi-carousel';
 import 'react-multi-carousel/lib/styles.css';
 import { LeftArrowIcon, RightArrowIcon } from '../Shared/SvgIcons';
@@ -11,20 +11,7 @@ import {
   CarouselButtonGroup,
   ArrowButton
 } from './styles';
-
-const communityList = [
-  { name: 'Ultron Apes', photo: '/assets/img/home/collection1.png', price: 1891000 },
-  { name: 'Ultron Apes', photo: '/assets/img/home/collection1.png', price: 1991000 },
-  { name: 'Ultron Apes', photo: '/assets/img/home/collection1.png', price: 1991000 },
-  { name: 'Ultron Apes', photo: '/assets/img/home/collection1.png', price: 1991000 },
-  { name: 'Ultron Apes', photo: '/assets/img/home/collection1.png', price: 1991000 },
-  { name: 'Ultron Apes', photo: '/assets/img/home/collection1.png', price: 1991000 },
-  { name: 'Ultron Apes', photo: '/assets/img/home/collection1.png', price: 1991000 },
-  { name: 'Ultron Apes', photo: '/assets/img/home/collection1.png', price: 1991000 },
-  { name: 'Ultron Apes', photo: '/assets/img/home/collection1.png', price: 1991000 },
-  { name: 'Ultron Apes', photo: '/assets/img/home/collection1.png', price: 1991000 },
-  { name: 'Ultron Apes', photo: '/assets/img/home/collection1.png', price: 1991000 },
-]
+import { getHistoricalCollections, getNearPrice } from '../../utils/paraApi';
 
 export const TopMarketCap = () => {
   const carouselRef = useRef<any>();
@@ -50,6 +37,24 @@ export const TopMarketCap = () => {
     }
   };
 
+  const [topMCAPCollections, setTopMCAPCollections] = useState<any[]>([]);
+
+  const getTopMCAPCollections = async () => {
+    const near_usd_value = await getNearPrice();
+    const collections = await getHistoricalCollections();
+    const derivatives = collections.sort((a: any, b: any) => b.est_market_cap - a.est_market_cap).slice(0, 10).map((data: any) => {
+      const { collection_info, est_market_cap } = data;
+      const { collection_name, media, discord, twitter, website } = collection_info;
+      return {
+        name: collection_name,
+        photo: media,
+        social_media: { discord, twitter, website },
+        est_market_cap: est_market_cap * near_usd_value
+      }
+    })
+    setTopMCAPCollections(derivatives);
+  }
+
   const goToNext = () => {
     const nextSlide = carouselRef.current.state.currentSlide + 1;
     carouselRef.current.goToSlide(nextSlide)
@@ -60,12 +65,15 @@ export const TopMarketCap = () => {
     carouselRef.current.goToSlide(prevSlide)
   }
 
+  useEffect(() => {
+    getTopMCAPCollections();
+  }, [])
+
   return (
     <Container>
       <Header className='container'>
         <TitleTabWrapper>
           <h1>Top marketcap Projects</h1>
-          <button className='primary-btn'>See All</button>
         </TitleTabWrapper>
         <CarouselButtonGroup>
           <ArrowButton onClick={gotToPrev}>
@@ -94,7 +102,7 @@ export const TopMarketCap = () => {
             autoPlay={true}
             ssr={true}
           >
-            {communityList.map((item, i) => (
+            {topMCAPCollections.map((item, i) => (
               <SingleMarketCapCard key={i} card={item} />
             ))}
           </Carousel>
