+++ package.json
@@ -14,6 +14,7 @@
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
     "@mui/material": "^5.6.4",
+    "bignumber.js": "^9.0.2",
     "chart.js": "^3.7.1",
     "cors": "^2.8.5",
     "express": "^4.18.0",
