--- pages/index_v2.tsx
@@ -1,12 +1,0 @@
-import type { NextPage } from 'next';
-import { Home as HomeController } from '../components/Home';
-
-const Home: NextPage = () => {
-  return (
-    <>
-      <HomeController />
-    </>
-  )
-}
-
-export default Home;
